Updated naming convention in Gui and added normalize signal functionality
Updated the naming convention for widgets and boxes in the Gui.

Added normalize a signal functionality to the gui and updated the normalize a signal function in the fiberClass.py file so it is compatible with the gui
