--- visualization.py
@@ -1,18 +1,0 @@
-import io
-import param
-import panel as pn
-import ipywidgets as ipw
-import plotly.express as px
-# import fibergui(non-cla
-from FiberClass import fiberObj
-
-
-def plot_raw_sig(event):
-    plot = obj.raw_signal_trace()
-    plot_pane = pn.pane.Plotly(height = 400, sizing_mode = "stretch_width")
-
-
-    template.main.append()
-                
-                
-
